+++ src/sessions.h
@@ -21,9 +21,9 @@ constexpr size_t message_n_label_bytes =
 struct Session {
   Session() = default;
   Session(size_t b, Sampling_status stat, uint32_t s, uint32_t d, uint16_t sp,
-          uint16_t dp, uint8_t p, std::vector<unsigned char> dat)
+          uint16_t dp, uint8_t p, uint64_t eid, std::vector<unsigned char> dat)
       : bytes_sampled(b), status(stat), src(s), dst(d), sport(sp), dport(dp),
-        proto(p), data(std::move(dat))
+        proto(p), session_event_id(eid), data(std::move(dat))
   {
   }
   Session(const Session& other) = default;
@@ -36,6 +36,7 @@ struct Session {
   uint16_t sport;
   uint16_t dport;
   uint8_t proto;
+  uint64_t session_event_id;
   std::vector<unsigned char> data;
 };
 
@@ -45,7 +46,8 @@ public:
   size_t get_number_bytes_in_sessions() const { return message_data; }
   size_t make_next_message(PackMsg& msg, uint64_t event_id);
   bool update_session(uint32_t src, uint32_t dst, uint8_t proto, uint16_t sport,
-                      uint16_t dport, const char* data, size_t len);
+                      uint16_t dport, const char* data, size_t len,
+                      uint64_t event_id);
   void set_allowed_entropy_ratio(float e) { entropy_ratio = e; }
   size_t size() const { return session_map.size(); }